[refactoring] Change implementation to Groovy
This completely removes Scala implementation classes in favor of Groovy.
Although it was fun to develop the plugin using Scala, Gradle is so much
Groovy friendly than Scala.

Now there is no need to keep an ever growing package.scala file with
tons of value classes to give Scala a nicer Gradle DSL.

This Groovy implementation attempts to make things at compile time as
much as possible using the @CompileStatic annotation whenever it is
viable.